- Use float to convert distro version   + The version my be denoted in decimal notation, such as '15.1'. This     triggers an exception during conversion when using int()

+++ azurelinuxagent/pa/rdma/factory.py
@@ -32,7 +32,7 @@ def get_rdma_handler(
     if (
             (distro_full_name == 'SUSE Linux Enterprise Server' or
              distro_full_name == 'SLES') and
-            int(distro_version) > 11
+            float(distro_version) > 11
     ):
         return SUSERDMAHandler()
